requirements.txt
asttokens==3.0.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.3
coverage==7.10.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
fonttools==4.59.0
git-filter-repo==2.47.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.5
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.3.2
packaging==24.2
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==308
pyzmq==26.2.1
requests==2.32.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
uv==0.7.6
wcwidth==0.2.13
xgboost==3.0.3
